add binary_file option to client request method Prior to this change, data sent was automatically sent as json and you couldn't send a binary file as data
This change adds the binary_file parameter which then sends it as data to requests.
This is extremely confusing since data in the wrapper is set to json while binary_file sends it as parameter data in requests.
If possible this should be changed but to keep backwards compatibility I will keep it as is.